+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kC/Dmc4kCAdapterSettings.cs
@@ -1,9 +1,10 @@
-﻿using ICD.Connect.Settings.Attributes;
+﻿using ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kCBase;
+using ICD.Connect.Settings.Attributes;
 
 namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kC
 {
 	[KrangSettings("Dmc4kC", typeof(Dmc4kCAdapter))]
-	public sealed class Dmc4kCAdapterSettings : AbstractInputCardSettings
+	public sealed class Dmc4kCAdapterSettings : AbstractDmc4kCBaseAdapterSettings
 	{
 	}
 }