+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kHdBase/AbstractDmc4kHdBaseAdapterSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHdBase
+{
+	public abstract class AbstractDmc4kHdBaseAdapterSettings : AbstractInputCardSettings
+	{
+	}
+}