+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kHdDspBase/AbstractDmc4kHdDspBaseAdapterRoutingControl.cs
@@ -1,0 +1,18 @@
+﻿#if SIMPLSHARP
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHdDspBase
+{
+	public abstract class AbstractDmc4kHdDspBaseAdapterRoutingControl<TParent> : AbstractInputCardAdapterRoutingControl<TParent>
+		where TParent : IInputCardAdapter
+	{
+		/// <summary>
+		/// Constructor.
+		/// </summary>
+		/// <param name="parent"></param>
+		/// <param name="id"></param>
+		protected AbstractDmc4kHdDspBaseAdapterRoutingControl(TParent parent, int id)
+			: base(parent, id)
+		{
+		}
+	}
+}
+#endif