+++ ICD.Connect.Routing.CrestronPro/Cards/Inputs/Dmc4kzHd/Dmc4kzHdAdapter.cs
@@ -1,21 +1,21 @@
-﻿using ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kCBase;
+﻿using ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kHdBase;
 #if SIMPLSHARP
 using Crestron.SimplSharpPro;
 using Crestron.SimplSharpPro.DM;
 #endif
 
-namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kC
+namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kzHd
 {
 #if SIMPLSHARP
-	public sealed class Dmc4kCAdapter :
-		AbstractDmc4kCBaseAdapter<Crestron.SimplSharpPro.DM.Cards.Dmc4kC, Dmc4kCAdapterSettings>
+	public sealed class Dmc4kzHdAdapter :
+		AbstractDmc4kHdBaseAdapter<Crestron.SimplSharpPro.DM.Cards.Dmc4kzHd, Dmc4kzHdAdapterSettings>
 	{
 		/// <summary>
 		/// Constructor.
 		/// </summary>
-		public Dmc4kCAdapter()
+		public Dmc4kzHdAdapter()
 		{
-			Controls.Add(new Dmc4kCAdapterRoutingControl(this, 0));
+			Controls.Add(new Dmc4kzHdAdapterRoutingControl(this, 0));
 		}
 
 		protected override bool GetIsOnlineStatus()
@@ -31,10 +31,10 @@ namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kC
 		/// <param name="cresnetId"></param>
 		/// <param name="controlSystem"></param>
 		/// <returns></returns>
-		protected override Crestron.SimplSharpPro.DM.Cards.Dmc4kC InstantiateCardExternal(byte cresnetId,
+		protected override Crestron.SimplSharpPro.DM.Cards.Dmc4kzHd InstantiateCardExternal(byte cresnetId,
 		                                                                                  CrestronControlSystem controlSystem)
 		{
-			return new Crestron.SimplSharpPro.DM.Cards.Dmc4kC(cresnetId, controlSystem);
+			return new Crestron.SimplSharpPro.DM.Cards.Dmc4kzHd(cresnetId, controlSystem);
 		}
 
 		/// <summary>
@@ -43,13 +43,13 @@ namespace ICD.Connect.Routing.CrestronPro.Cards.Inputs.Dmc4kC
 		/// <param name="cardNumber"></param>
 		/// <param name="switcher"></param>
 		/// <returns></returns>
-		protected override Crestron.SimplSharpPro.DM.Cards.Dmc4kC InstantiateCardInternal(uint cardNumber, Switch switcher)
+		protected override Crestron.SimplSharpPro.DM.Cards.Dmc4kzHd InstantiateCardInternal(uint cardNumber, Switch switcher)
 		{
-			return new Crestron.SimplSharpPro.DM.Cards.Dmc4kC(cardNumber, switcher);
+			return new Crestron.SimplSharpPro.DM.Cards.Dmc4kzHd(cardNumber, switcher);
 		}
 	}
 #else
-	public sealed class Dmc4kCAdapter : AbstractDmc4kCBaseAdapter<Dmc4kCAdapterSettings>
+	public sealed class Dmc4kzHdAdapter : AbstractDmc4kHdBaseAdapter<Dmc4kzHdAdapterSettings>
 	{
 	}
 #endif